Always update internal switch state from MQTT
Previously the internal state is only updated when a switch action
occurs (i.e. when the user physically operates the switch), and
programmatic switch actions are ignored and thus unknown to regelwerk.

This sometimes led to regelwerk not responding properly until the switch
is physically pressed to force an update to the internal state.