Created a more generic framework
Created the concept of tracked Devices and Timers. Timers are now
(explicitly) used as the mechanism for session management. When the door
is opened and conditions are right, a Timer is created to signify a
"session". It is then stopped and started as necessary until it finally
fires to automatically turn off the light, then it is destroyed.

Devices help to keep track of state to simplify handler code. You can
either choose to receive all device events, or only when the device
state changes. The code currently uses both types.